### 2024-03-03 #6268 @arfon
@JoshSchramm94 – one minor issue while we wait here, this paper is rather longer than we usually accept in JOSS. Ideally, we're looking for papers ~1000 words in length. I'm not sure if there are obvious ways in which you might be able to shorten this (e.g., move some of the examples into docs) but I just wanted to mention this before we get too far into the review process.
