### 2024-03-15 #6268 @arfon
@RMeli – would you be willing to edit this submission for JOSS?
