### 2024-05-01 #6708 @editorialbot
[REVIEW]: validateHOT
**Submitting author:** @JoshSchramm94 (<a href="http://orcid.org/0000-0001-5602-4632">Joshua Benjamin Schramm</a>)
**Repository:** https://github.com/JoshSchramm94/validateHOT
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.5
