### 2024-06-13 d27d06d
two syntax changes
- use seq_* functions instead of 1: for safety
- use && & || for scalar conditionals

both are suggestions, tests passed.

I'm curious, is there any particular reason you use base:: for base R functions?
