+++ R/zc_diffs.R
@@ -154,7 +154,7 @@ zc_diffs <- function(data, group = NULL, attrib, coding, interpolate.levels = NU
 
   # test numeric input for coding
   if (!(base::is.null(coding))) {
-    for (i in 1:base::length(coding)) {
+    for (i in base::seq_along(coding)) {
       if (!(base::is.numeric(coding[i]))) {
         base::stop("Error: 'coding' only can have numeric input!")
       }
@@ -170,14 +170,14 @@ zc_diffs <- function(data, group = NULL, attrib, coding, interpolate.levels = NU
 
 
   # test input of interpolate levels
-  if (!(base::is.list(interpolate.levels)) &
+  if (!(base::is.list(interpolate.levels)) &&
     !(base::is.null(interpolate.levels))) {
     base::stop("Error: Input of 'interpolate.levels' has to be a list!")
   }
 
   # test variables of interpolate.levels
   if (!(base::is.null(interpolate.levels))) {
-    for (tt in 1:base::length(interpolate.levels)) {
+    for (tt in base::seq_along(interpolate.levels)) {
       lng <- base::length(interpolate.levels[[tt]])
 
       for (lng_lev in 1:lng) {
@@ -192,7 +192,7 @@ zc_diffs <- function(data, group = NULL, attrib, coding, interpolate.levels = NU
   }
 
   # interpolate.levels can not be larger than number of attributes
-  if (!base::is.null(interpolate.levels) &
+  if (!base::is.null(interpolate.levels) &&
     (base::length(interpolate.levels) > base::length(attrib))) {
     base::stop(
       "Error: List of 'interpolate.levels' can not be larger than list of 'attrib'!"
@@ -202,20 +202,20 @@ zc_diffs <- function(data, group = NULL, attrib, coding, interpolate.levels = NU
   # check length of 'attrib' and coding
   # test input of list in prod.levels
   if (!(base::is.null(attrib))) {
-    for (tt in 1:base::length(attrib)) {
-      if (base::length(attrib[[tt]]) == 1 & coding[tt] == 0) {
+    for (tt in base::seq_along(attrib)) {
+      if (base::length(attrib[[tt]]) == 1 && coding[tt] == 0) {
         base::stop(
           "Error: If attribute is part-worth coded at least 2 attribute levels need to be specified!"
         )
       }
 
-      if (base::length(attrib[[tt]]) > 1 & coding[tt] == 1) {
+      if (base::length(attrib[[tt]]) > 1 && coding[tt] == 1) {
         base::stop(
           "Error: If attribute is linear coded only one attribute level needs to be specified!"
         )
       }
 
-      if (base::length(attrib[[tt]]) == 1 & coding[tt] == 2) {
+      if (base::length(attrib[[tt]]) == 1 && coding[tt] == 2) {
         base::stop(
           "Error: If attribute is piecewise coded at least 2 attribute levels need to be specified!"
         )
@@ -224,14 +224,14 @@ zc_diffs <- function(data, group = NULL, attrib, coding, interpolate.levels = NU
   }
 
   # if 1 is coded, interpolate.levels needs to be specified
-  if (base::any(coding == 1) & base::is.null(interpolate.levels)) {
+  if (base::any(coding == 1) && base::is.null(interpolate.levels)) {
     base::stop(
       "Error: 'interpolate.levels' is missing!"
     )
   }
 
   # if no 1 coded interpolate.levels is not needed
-  if (!base::any(coding == 1) & !base::is.null(interpolate.levels)) {
+  if (!base::any(coding == 1) && !base::is.null(interpolate.levels)) {
     base::stop(
[…]
