### 2024-06-13 9a1516f
Example in README checks out
- minor changes, not sure about column index change

openjournals/joss-reviews/issues/6708
