+++ README.Rmd
@@ -366,7 +366,7 @@ and as well as the `id`. Since we also have a *no-buy* alternative in
 our validation task, we specify the `none` argument, otherwise we would
 have left it empty. Afterwards, we define each alternative with the
 argument `prod.levels`. If we look back at the data frame, we can see
-that the first alternative in the holdout task (`c(4, 9, 19)`) is
+that the first alternative in the holdout task (`c(4, 9, 18)`) is
 composed of the following attribute levels `Att1_Lev1`, `Att2_Lev1`, and
 `Att3_Lev5`.
 
@@ -544,7 +544,7 @@ hitrate(
 
 In this case, the group assignment is stored in form of integers.
 However, the output is the same if it is a factor or labelled data. To
-proof this we just quickly change `group` into a factor by using the
+prove this we just quickly change `group` into a factor by using the
 `factor()` function provided by the *R* Core Team (2023).
 
 ```{r}
