### 2024-06-13 JoshSchramm94/validateHOT#3 @jamesuanhoro
Additional issues
- [x] README references a vignette, I don't find it when I run `vignette(package = "validateHOT")`, appears to be WIP
- [x] Provide a means for community members to contribute to the development of the package, this is required by JOSS
- [x] The typical workflow section in the paper should show the required packages for the code in the paper. One can infer they need to load `validateHOT`, but inexperienced R users may not realize they also need `dplyr`.

openjournals/joss-reviews/issues/6708
