+++ README.Rmd
@@ -69,7 +69,7 @@ your validation/holdout tasks you also need to provide the actual choice
 of your validation/holdout task. You can use the `merge()` function
 provided by the *R* base package to do so (R Core Team, 2023). We
 provide a short tutorial in the markdown, for a more detailed tutorial,
-please see the vignette that comes along with `validateHOT`.
+please see the vignette that comes along with `validateHOT` (`vignette("validateHOT", package = "validateHOT")`).
 
 👈🏾 <u>**What you get**</u>:<br> At the moment, `validateHOT` provides
 functions for 4 key components:
