### 2024-07-03 d19b19b
updated tests and also changed version to 1.0.1 - 240703
