+++ paper/paper.bib
@@ -321,3 +321,11 @@ analysis of Best-Worst Surveys},
 	url = {http://dx.doi.org/10.1007/s11747-023-00997-5},
 	langid = {en}
 }
+
+@article{dplyr,
+	title = {dplyr: A Grammar of Data Manipulation},
+	author = {Wickham, Hadley and {François}, Romain and Henry, Lionel and {Müller}, Kirill and Vaughan, Davis},
+	year = {2023},
+	date = {2023},
+	url = {https://CRAN.R-project.org/package=dplyr}
+}
