+++ paper/paper.Rmd
@@ -19,20 +19,15 @@ affiliations:
  - name: Otto von Guericke University of Magdeburg, Germany
    index: 1
 citation_author: Schramm & Lichters
-date: 04 March
+date: 07 July
 year: 2024
 bibliography: paper.bib
 link-citations: true
-output: rticles::joss_article
-journal: JOSS
-header-includes: 
- - \usepackage{float}
- - \usepackage{fancyhdr}
+output: pdf_document
 ---
 
-```{=tex}
 \newcommand{\colcod}[1]{\texttt{\color{purple}#1}}
-```
+
 ```{r include=FALSE}
 library(validateHOT)
 library(tidyverse)
@@ -53,19 +48,17 @@ validateHOT combines both validation and market simulation in one package and ha
 Other packages provide functions to calculate validation metrics, however, these are not always specified for individual part-worth utilities. Metrics [@Metrics], for example, provide functions to run validation metrics such as *mean absolute error*, *root mean squared error*, or the five metrics of the confusion matrix. However, to get the output of, for example, Sawtooth Software or ChoiceModelR [@ChoiceModelR] into the right format, the user needs some data wrangling. The package conjoint [@conjoint] provides functions that are most similar to validateHOT's ones, but no validation functions are included and the package focuses on classical conjoint analysis, thus it is limited when applying more common conjoint methods. logitr [@logitr] provides market simulations tools, however, no validation metrics such as mean hit probability or hit rate. \autoref{comparison} shows a comparison of validateHOT's functions with current R packages. To the best of our knowledge, a package that converts raw utility scores into validation metrics or running a variety of marketing simulations (especially TURF) is missing.
 
 ```{=tex}
-\begin{figure}[h]
+\begin{figure}[ht]
   \includegraphics{figures/FunctionComparison.png}
   \caption{Comparison of validateHOT's function to existing R packages}
   \label{comparison}
 \end{figure}
 ```
-
 validateHOT is introduced with data estimated with Lighthouse Studio using effects-coding for creating the design matrix. It, however, can easily be used with data estimated with ChoiceModelR [@ChoiceModelR], bayesm [@bayesm], or STAN [@rstan], if used with similar settings (ChoiceModelR, for example, automatically implements effects-coding).
 
 # Key functions
 
 validateHOT's functions can be categorized into four main components, see \autoref{tab:table1}. To bring the data into the right format for some functions, the \texttt{\color{purple}createHOT()} function can be applied, which creates each alternatives' total utility by applying the additive utility model.
-
 
 | Validation metrics | Confusion matrix | Market simulations | Rescaling scores |
 |:----------------:|:----------------:|:----------------:|:----------------:|
@@ -80,14 +73,13 @@ validateHOT's functions can be categorized into four main components, see \autor
 
 # Typical workflow
 
-We provide the workflow for a MaxDiff study and a CBC study with only part-worth coded attributes (the vignette provides detailed examples for other CBCs and an ACBC).
+We provide the workflow for a MaxDiff study and a CBC study with only part-worth coded attributes (the vignette provides detailed examples for other CBCs and an ACBC). To run the following code chunks, make sure to also install the dplyr package [@dplyr] since it is a dependency of validateHOT.
 
 ## MaxDiff
 
 ### Creating Holdout Task / Market Scenario
 
-After running the HB estimation [@allenby1995], the **raw** utility scores have to be exported and read into an *R* data frame. 
[…]
