+++ vignettes/references.bib
@@ -1,0 +1,9 @@
+
+@article{remotes,
+	title = {remotes: R Package Installation from Remote Repositories, Including
+'GitHub'},
+	author = {{Csárdi}, {Gábor} and Hester, Jim and Wickham, Hadley and Chang, Winston and Morgan, Martin and Tenenbaum, Dan},
+	year = {2024},
+	date = {2024},
+	url = {https://CRAN.R-project.org/package=remotes}
+}
