### 2024-07-04 3782794
added code of conduct and contributing file - 240704
