+++ contributing.md
@@ -1,0 +1,58 @@
+# Contributing
+
+Contributions to the validateHOT project are welcomed. For an easy contribution process, please carefully read and follow the guidelines.
+
+## Quicklinks
+
+* [Code of Conduct](#code-of-conduct)
+* [Getting Started](#getting-started)
+    * [Issues](#issues)
+    * [Pull Requests](#pull-requests)
+* [Getting Help](#getting-help)
+* [Attribution](#attribution)
+
+## Code of Conduct
+
+We take our open source community seriously and hold ourselves and other contributors to high standards of communication. By participating and contributing to this project, you agree to uphold our [Code of Conduct]().
+
+## Getting Started
+
+Contributions can be made via Issues and Pull Requests (PRs). A few general guidelines cover both:
+
+- Before creating your own Issue, please carefully look through the existing one to check whether your Issue is already adressed.
+- We try to address issues as soon as possible. However, depending on the problem and maintainer availability, it could take a while to investigate the problem. In case an issue has not received any attention for a while, please send a friendly reminder. 
+
+### Issues
+
+Issues should be used to report problems with the library, request a new feature, or to discuss potential changes before a PR is created. 
+Please provide exemplarily code to when reporting the bug to better understand the problem. 
+
+If you find an Issue that addresses the problem you're having, please add your own reproduction information to the existing issue rather than creating a new one. Adding a [reaction](https://github.blog/2016-03-10-add-reactions-to-pull-requests-issues-and-comments/) can also help be indicating to our maintainers that a particular problem is affecting more than just the reporter.
+
+### Pull Requests
+
+PRs are always welcome and can be a quick way to get your fix or improvement slated for the next release. However, please always open an Issue before submitting a PR.
+
+In general, PRs should:
+
+- Only fix/add the functionality in question **OR** address wide-spread whitespace/style issues, not both.
+- Add unit or integration tests for fixed or changed functionality (if a test suite already exists).
+- Address a single concern in the least number of changed lines as possible.
+
+
+In general, we follow the [GitHub flow](https://guides.github.com/introduction/flow/index.html) development model:
+
+1. Fork the repository to your own Github account
+2. Clone the project to your machine
+3. Create a branch locally with a succinct but descriptive name
+4. Commit changes to the branch
+5. Push changes to your fork
+6. Open a PR in our repository and follow the PR template so that we can efficiently review the changes.
+
+## Getting Help
+
+Please use the repository's [Issue section](https://github.com/JoshSchramm94/validateHOT/issues).
+
+## Attribution
+
+This document was adapted from the [General Contributing Guidelines](https://github.com/auth0/open-source-template/blob/master/GENERAL-CONTRIBUTING.md) of the auth0 project.
