### 2024-07-04 e818755
Update contributing.md
added url for code of conduct
