+++ DESCRIPTION
@@ -22,21 +22,21 @@ BugReports: https://github.com/JoshSchramm94/validateHOT/issues
 Depends:
     R (>= 3.5.0)
 Imports: 
-    dplyr (>= 1.1.0),
+    dplyr (>= 1.1.4),
     fastDummies (>= 1.7.3),
     magrittr (>= 2.0.3),
-    scales (>= 1.2.1),
-    stats (>= 4.2.2),
+    scales (>= 1.3.0),
+    stats (>= 4.3.2),
     tibble (>= 3.2.1),
-    tidyr (>= 1.3.0),
-    tidyselect (>= 1.2.0),
-    utils (>= 4.2.3)
+    tidyr (>= 1.3.1),
+    tidyselect (>= 1.2.1),
+    utils (>= 4.3.2)
 Suggests: 
-    knitr (>= 1.43),
-    labelled (>= 2.12.0),
+    knitr (>= 1.47),
+    labelled (>= 2.13.0),
     Metrics (>= 0.1.4),
-    rmarkdown (>= 2.24),
-    testthat (>= 3.0.0)
+    rmarkdown (>= 2.27),
+    testthat (>= 3.2.1.1)
 VignetteBuilder: 
     knitr,
     dplyr,
@@ -44,4 +44,4 @@ VignetteBuilder:
 Config/testthat/edition: 3
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.1
