### 2024-07-04 JoshSchramm94/validateHOT#3 @JoshSchramm94
@jamesuanhoro, thank you very much for your valuable feedback as well as taking your time to review the paper and the package. I now updated everything and also added a contributing file as well as a conde of conduct. I also updated the workflow in the paper to ensure that readers are aware they need the `dplyr` package as well. Once again thanks for your feedback.
