### 2024-08-22 f1d132a
updated tidyverse logic in functions - 240822
