### 2024-08-22 4287dd8
added news.md - 240822
