+++ NEWS.md
@@ -1,0 +1,4 @@
+# validateHOT 1.0.2
+
+* Updated tidyverse logic and switched from summarise() to reframe()
+* updated prob_scores() function for individual probability scores
