### 2024-10-28 9172ae8
started goodpractice() - 241028
