+++ R/ACBC_interpolate.R
@@ -1,4 +1,5 @@
-#' ACBC data set with part-worth coded attributes, linear-coded attribute and piecewise-coded price
+#' ACBC data set with part-worth coded attributes, linear-coded attribute and
+#' piecewise-coded price
 #'
 #' A data set with raw scores export for an adaptive choice-based conjoint
 #' including a linear-coded attribute as well as piecewise-coded price.
