+++ R/att_imp.R
@@ -8,25 +8,26 @@
 #' coding, '1' = linear coding, or '2' = piecewise coding.
 #' @param interpolate.levels A list of the attribute levels that should
 #' be interpolated. These have to be the same as specified in model estimation
-#' (e.g., if you center attribute levels before estimation, insert the centered levels).
-#' Please make sure to provide the whole list. Only has to be specified for the
-#' variables that are coded as '1' (linear).
+#' (e.g., if you center attribute levels before estimation, insert the centered
+#' levels). Please make sure to provide the whole list. Only has to be
+#' specified for the variables that are coded as '1' (linear).
 #' @param res A vector indicating whether individual shares (\code{ind}) or
 #' aggregated (\code{agg}) shares should be returned.
 #'
 #' @details
-#' \code{att_imp} converts raw utilities of a CBC or an ACBC to relative importance
-#' scores (see, Orme, 2020, p. 80, for more information).
+#' \code{att_imp} converts raw utilities of a CBC or an ACBC to relative
+#' importance scores (see, Orme, 2020, p. 80, for more information).
 #'
-#' \code{data} has to be a data frame with all attributes and the corresponding levels.
-#' Attribute levels need to be the raw utilities of hierarchical Bayes estimation.
+#' \code{data} has to be a data frame with all attributes and the corresponding
+#' levels. Attribute levels need to be the raw utilities of hierarchical Bayes
+#'  estimation.
 #'
 #' \code{group} optional grouping variable, if results should be displayed by
 #' different groups. Has to be column name of variables in \code{data}.
 #'
 #' \code{attrib} specifies the attribute levels for each alternative.
-#' Input for \code{attrib} has to be a list. Needs to specify the column names or
-#' column indexes of the attribute levels.
+#' Input for \code{attrib} has to be a list. Needs to specify the column names
+#' or column indexes of the attribute levels.
 #'
 #' \code{coding} is required to indicate the attribute coding. \code{0}
 #' to indicated part-worth coding, \code{1} for linear coding, or \code{2} for
@@ -38,14 +39,15 @@
 #' All values have to be specified. For example, if one linear coded attribute
 #' has 5 levels, all 5 levels have to be inserted.
 #'
-#' \code{res} specifies whether results should be aggregated across all participants
-#' or across \code{group} (\code{res} needs to be set to \code{agg}) or if scores
-#' should be converted for individuals only.
+#' \code{res} specifies whether results should be aggregated across all
+#' participants or across \code{group} (\code{res} needs to be set to
+#' \code{agg}) or if scores should be converted for individuals only.
 #'
 #'
 #' @seealso {
 #' \code{\link[=prob_scores]{prob_scores}} for probability scores for MaxDiff
-#' \code{\link[=zero_anchored]{zero_anchored}} for zero-anchored interval scores for MaxDiff
+#' \code{\link[=zero_anchored]{zero_anchored}} for zero-anchored interval
+#' scores for MaxDiff
 #' \code{\link[=zc_diffs]{zc_diffs}} for zero-centered diff scores for (A)CBC
 #' }
 #'
@@ -58,8 +60,9 @@
 #'
 #' @references {
 #'
-#' Orme, B. K. (2020). \emph{Getting Started with Conjoint Analysis: Strategies for Product Design and Pricing Research}.
[…]
