+++ tests/testthat/test-ACBC_interpolate.R
@@ -7,7 +7,9 @@ test_that("Structure of DF", {
 })
 
 test_that("Variables numeric", {
-  names <- base::colnames(ACBC_interpolate)[-(base::which(base::colnames(ACBC_interpolate) == "Group"))]
+  names <- base::colnames(
+    ACBC_interpolate)[-(base::which(
+      base::colnames(ACBC_interpolate) == "Group"))]
   for (i in base::seq_along(names)) {
     expect_true(!base::is.character(ACBC_interpolate[names[i]]))
   }
