+++ R/createHOT.R
@@ -4,25 +4,25 @@
 #' for the alternatives in a market scenario.
 #'
 #' @param data A data frame with all relevant variables.
-#' @param id A vector of unique identifier in \code{data}.
-#' @param none An optional vector to specify \code{none}
-#' alternative in \code{data}.
+#' @param id A vector of unique identifier in `data`.
+#' @param none An optional vector to specify `none`
+#' alternative in `data`.
 #' @param prod.levels A list to define the attribute levels of the
-#' alternatives (\code{prod}). If linear-coded or piecewise-coded
+#' alternatives (`prod`). If linear-coded or piecewise-coded
 #' attributes are included, column indexes are required for the input.
 #' @param interpolate.levels A list of the levels of the attribute that should
 #' be interpolated. These have to be the same as specified in model estimation
 #' (e.g., if you scale or center attribute levels before estimation, insert the same levels).
 #' Please make sure to provide the whole list. Only has to be specified for the
-#' attributes that are coded as '1' (linear) or '2' (piecewise).
+#' attributes that are coded as `1` (linear) or `2` (piecewise).
 #' @param piece.p A list of the lower level and the upper
 #' level that should be used for interpolating.
 #' @param lin.p A vector to specify linear coded variables.
-#' @param coding A vector to define attributes coding, '0' = part-worth
-#' coding,'1' = linear coding, '2' = piecewise coding; please make sure to code
+#' @param coding A vector to define attributes coding, `0` = part-worth
+#' coding,`1` = linear coding, `2` = piecewise coding; please make sure to code
 #' linear price of ACBC as piecewise since there are usually two values to interpolate.
-#' @param method A character to specify the \code{method} of your study.
-#' \code{method} has to be one of the following: "MaxDiff", "CBC", or "ACBC".
+#' @param method A character to specify the `method` of your study.
+#' `method` has to be one of the following: `"MaxDiff"`, `"CBC"`, or `"ACBC"`.
 #' @param varskeep A vector specifying variables that should be kept
 #' in the data frame.
 #' @param choice Actual choice in the holdout/validation task. Leave empty for
@@ -32,74 +32,74 @@
 #' To test the validation metrics of a holdout/validation task or to run a
 #' market scenario, the scenario first has to be created, by summing up
 #' the alternatives part-worth utilities.
-#' This is done with the \code{createHOT} function.
+#' This is done with the `createHOT()` function.
 #' Make sure you upload the raw utilities of your study (either from Sawtooth Software
 #' or ChoiceModelR, Sermas, 2022).
 #' The function then creates the alternatives' utility based on the additive
 #' utility model (Rao, 2014, p. 82). If you are working with alternative specific-designs,
-#' insert \code{NA} if attribute is not specified.
+#' insert `NA` if attribute is not specified.
 #'
-#' \code{data} has to be a data frame with raw scores of the attribute
+#' `data` has to be a data frame with raw scores of the attribute
 #' levels.
 #'
-#' \code{id} has to be the column index or column name of the id (unique for each participant)
+#' `id` has to be the column index or column name of the id (unique for each participant)
 #' in data frame.
 #'
-#' \code{none} is required in case a \code{none} alternative is
+#' `none` is required in case a `none` alternative is
 #' included in holdout/validation task, specify
[…]
