+++ R/Recall.R
@@ -1,12 +1,13 @@
 #' Function to calculate recall of confusion matrix for general product demand
 #'
-#' @description \code{recall} is one of the 5 metrics of the confusion matrix
+#' @description
+#' `recall()` is one of the 5 metrics of the confusion matrix
 #' and is defined as \eqn{\frac{TP}{TP + FN}}, where TP =
 #' True Positives, FN = False Negatives (see, e.g., Burger, 2018).
 #'
 #' @param data A data frame with all relevant variables.
 #' @param group Optional column name(s) to specify grouping variable(s)
-#' to get \code{recall} by group(s).
+#' to get `recall()` by group(s).
 #' @param opts Column names of the alternatives included in the
 #' validation/holdout task.
 #' @param choice Column name of the actual choice in the validation/holdout
@@ -14,31 +15,31 @@
 #' @param none Column name of none alternative.
 #'
 #' @details
-#' The current logic of \code{recall} is to determine whether a binary coded
+#' The current logic of `recall()` is to determine whether a binary coded
 #' choice is correctly predicted by the model. To use the function, the
-#' validation/holdout task must include a \code{none} alternative.
+#' validation/holdout task must include a `none` alternative.
 #' One possible application is, for example, whether a buy or a no-buy choice
 #' has been correctly predicted. For example, suppose you have three
-#' alternatives plus a \code{none} alternative and want to check whether a buy
+#' alternatives plus a `none` alternative and want to check whether a buy
 #' or no-buy was correctly predicted. This function can be useful if you, for
 #' example, test whether your model significantly overestimates or
 #' underestimates a purchase likelihood.
 #'
-#' \code{data} has to be a data frame including the alternatives shown in
-#' the validation/holdout task. Can be created using the \code{createHOT()}
+#' `data` has to be a data frame including the alternatives shown in
+#' the validation/holdout task. Can be created using the `createHOT()`
 #' function.
 #'
-#' \code{group} optional grouping variable(s), if results should be displayed
-#' by different group(s). Has to be column name of variables in \code{data}.
+#' `group` optional grouping variable(s), if results should be displayed
+#' by different group(s). Has to be column name of variables in `data`.
 #'
-#' \code{opts} is required to specify the different alternatives in the
-#' validation/holdout task (also includes the \code{none} alternative).
-#' Input of \code{opts} has to be column names of variables in \code{data}.
+#' `opts` is required to specify the different alternatives in the
+#' validation/holdout task (also includes the `none` alternative).
+#' Input of `opts` has to be column names of variables in `data`.
 #'
-#' \code{choice} to specify column of actual choice in the validation/holdout
-#' task. Input of \code{choice} has to be column name of actual choice.
+#' `choice` to specify column of actual choice in the validation/holdout
+#' task. Input of `choice` has to be column name of actual choice.
 #'
-#' \code{none} is required to specify column name of the \code{none}
+#' `none` is required to specify column name of the `none`
 #' alternative in the validation/holdout task.
 #'
 #' Please be aware about the following 2x2 table regarding coding of buy and
@@ -105,90 +106,92 @@
 
[…]
