### 2024-11-29 #6708 @njtierney
Hi @skanwal ! Thanks for your patience here, here is my review - let me know if I can provide any more information?

# Functionality: Have the functional claims of the software been confirmed?

## 1 test failure

Firstly I'd like to thank the authors for writing tests, well done! Currently, this package does not currently pass checks from `devtools::check()`, this is due to one failing test:

```
── Failure ('test-turf.R:216:3'): if 'fc' 'reach' and 'freq'  ──────────────────
base::all(t2$reach == (t2$freq * 100)) is not TRUE

`actual`:   FALSE
`expected`: TRUE

[ FAIL 1 | WARN 0 | SKIP 0 | PASS 835 ]
Error: Test failures
Execution halted

1 error ✖ | 0 warnings ✔ | 0 notes ✔
Error: R CMD check found ERRORs
Execution halted

Exited with status 1.

```

## goodpractice checks

Running `goodpractice::gp()` I get the following recommendations to change, which I believe the author should change, and they should also run `goodpractice::gp()` themselves.

```
── GP validateHOT ───────────────────────────────

It is good practice to

  ✖ write short and simple
    functions. These functions have high
    cyclomatic complexity (>50): createHOT
    (116), turf (60), zc_diffs (54). You
    can make them easier to reason about by
    encapsulating distinct steps of your
    function into subfunctions.
  ✖ avoid long code lines, it is
    bad for readability. Also, many people
    prefer editor windows that are about 80
    characters wide. Try make your lines
    shorter than 80 characters

    R/ACBC_interpolate.R:1:81
    R/Accuracy.R:15:81
    R/Accuracy.R:19:81
    R/Accuracy.R:20:81
    R/Accuracy.R:23:81
    ... and 677 more lines

  ✖ avoid 1:length(...),
    1:nrow(...), 1:ncol(...), 1:NROW(...)
    and 1:NCOL(...) expressions. They are
    error prone and result 1:0 if the
    expression on the right hand side is
    zero. Use seq_len() or seq_along()
    instead.

    R/Accuracy.R:143:13
    R/att_imp.R:108:15
    R/att_imp.R:131:16
    R/att_imp.R:156:16
    R/att_imp.R:238:15
    ... and 61 more lines

  ✖ checking tests ... Running
    ‘testthat.R’ ERROR Running the tests in
    ‘tests/testthat.R’ failed. Last 13
    lines of output: int [1:70] 2 2 3 3 3 3
    3 3 1 1 ... chr [1:32] "Group 1" "Group
    1" "Group 1" "Group 1" "Group 1" "Group
    1" ... chr [1:70] "Group 1" "Group 2"
    "Group 1" "Group 2" "Group 1" "Group 2"
    ... [ FAIL 1 | WARN 0 | SKIP 0 | PASS
    835 ] ══ Failed tests
    ════════════════════════════════════════════════════════════════
    ── Failure ('test-turf.R:216:3'): if
    'fc' 'reach' and 'freq'
    ────────────────── base::all(t2$reach
    == (t2$freq * 100)) is not TRUE
    `actual`: FALSE `expected`: TRUE [ FAIL
    1 | WARN 0 | SKIP 0 | PASS 835 ] Error:
    Test failures Execution halted
  ✖ avoid 'T' and 'F', as they are
    just variables which are set to the
    logicals 'TRUE' and 'FALSE' by default,
    but are not reserved words and hence
    can be overwritten by the user.  Hence,
    one should always use 'TRUE' and
    'FALSE' for the logicals.

    R/KL.R:NA:NA
    R/KL.R:NA:NA
    R/MAE.R:NA:NA
    R/MarkSim.R:NA:NA
    R/MarkSim.R:NA:NA
    ... and 2 more lines

─────────────────────────────────────────────────
```

## Data provided with the package

Great to see data sets provided in the package, and especially that the authors have documented the variables. Some small suggestions:

- [ ] Provide some example code with it that demonstrates what to do or what is expected to be done with the data.
- [ ] Document the type of variable for each variable. For example
- [ ] use lowercase snake_case for variable names. e.g., `Price_1.99` should be `price_1_99`.
- [ ] Provide a source for the data. Is it simulated? It it pulled from somewhere in particular? If so, it should be cited.

I've got some specific comments on some of the data

- ACBC_interpolate.R
  - `Price_6.99` says "Piecewise coding for price at $5." Why is it not `Price_4.99` or Price_5.99 or similar?
  - Similarly for the other variable names, `Price_10.99. Piecewise coding for price at $15`; `Price_12.99. Piecewise coding for price at $20.`; `Price_15.99. Piecewise coding for price at $25.`; `Price_25.99. Piecewise coding for price at $27.`
- ACBC.R - all good.
- CBC_lin.R - all good.
- CBC.R - all good.
- MaxDiff.R - all good

Where are these data sources taken from? These files should ideally be a data-raw folder that accompanies these and shows how the data was created. See https://r-pkgs.org/data.html#sec-data-data-raw for more information on how to do this.

### Accuracy.R

Lines like

```r
#' @importFrom dplyr group_by summarise select pick
```

Can be replaced by putting `dplyr` in Imports in DESCRIPTION (with `usethis::use_package("dplyr")`). Using `@importFrom pkg fun` will mean that the package is loaded when the user loads the package - see https://r-pkgs.org/dependencies-in-practice.html#sec-dependencies-NAMESPACE-workflow for more detail.

In examples,

```r
#' library(validateHOT)
```

Should be removed, the package is already loaded.

The pattern:

```r
data %>%
    dplyr::select(., {{ opts }})
```

Is unneeded, since by default `data` will be the first argument of `select`. It is equivalent to write:

```r
data %>%
    dplyr::select({{ opts }})
```

You have written a lot of tests to check for missing values and the correct types in the data, which is very thorough defensive programming! Would you consider wrapping these up into a couple of functions, so that we can get closer to the main function of accuracy? I've called this type of pattern an "error handling eclipse", and wrote about it here: https://www.njtierney.com/post/2023/12/06/long-errors-smell/

I want to emphasise that it is great that you have written so many checks! But this is just a suggestion to help make your code easier to reuse for yourself. You've written out a lot of these checks throughout your package, which is great, but it also creates a bit of extra work to keep them all in check. Let me know if you'd like some help wrapping up these checks into sets of functions?

This pattern:

```r
 # check for wrong / missing input
  if (base::length(data %>% dplyr::select(., {{ none }})) == 0) {
    stop("Error: argument 'none' is missing!")
  }

  # check whether opts are defined
  if (base::length(data %>% dplyr::select(., {{ opts }})) == 0) {
    stop("Error: argument 'opts' is missing!")
  }
```

Is intended to find whether the argument "none" was provided, right? You can instead use `missing`, like so:

```r
if (missing(none)) {
    stop("Error: argument 'none' is missing!")
  }
```

The code below:

```r
  if (base::length(data %>% dplyr::select(., {{ opts }})) == 1) {
    stop("Error: specify at least 2 alternatives in 'opts'!")
  }
```

I believe should be rewritten as follows:

```r
opts_cols <- dplyr::select(data, {{ opts }})
n_opts_cols <- ncol(opts_cols)
if (n_opts_cols <= 1) {
    stop("Error: specify at least 2 alternatives in 'opts'!")
  }
```

This provides the if statement with an "explaining variable", which makes it
easier to understand what the longer statement does. In addition using `ncol` is more precise than using `length`, as length could imply to the reader that it is measuring the number of rows.

```r
if (!(data %>% dplyr::select(., {{ none }}) %>% base::colnames()) %in%
    (data %>% dplyr::select(., {{ opts }}) %>% base::colnames())) {
    stop("Error: 'none' has to be part of 'opts'!")
  }
```

Should be rewritten as:

```r
names_none <- dplyr::select(data, {{ none }}) %>% names()
names_opts <- dplyr::select(data, {{ opts }}) %>% names()
none_in_opts <- names_none %in% names_opts
if (!none_in_opts) {
    stop("Error: 'none' has to be part of 'opts'!")
  }
```

These checks for missing values could be wrapped up in a function:

```r
  # grouping variable
  ## check for missings
  if (base::anyNA(data %>% dplyr::select(., {{ group }}))) {
    warning("Warning: 'group' contains NAs!")
  }
```

```r
any_var_na <- function(data, selector) {
  selected_vars <- data %>%
    dplyr::select( {{ selector }})

  anyNA(selected_vars)
}

group_contains_na <- any_var_na(data, {{ group }} )
  # grouping variable
  ## check for missings
  if (group_contains_na) {
    warning("Warning: 'group' contains NAs!")
  }
```

The following can be rewritten to use the already existing `opts_cols` data

```r
  # alternatives
  ## store names of alternatives
  alternatives <- data %>%
    dplyr::select(., {{ opts }}) %>%
    base::colnames()

  ## check whether variable is numeric
  for (i in 1:base::length(alternatives)) {
    if (!base::is.numeric(data[[alternatives[i]]])) {
      stop("Error: 'opts' has to be numeric!")
    }
  }
```

As:

```r
  is_opts_numeric <- vapply(
    X = opts_cols,
    FUN = is.numeric,
    FUN.VALUE = TRUE
    )

  any_opts_not_numeric <- any(!is_opts_numeric)

  if (any_opts_not_numeric) {
      stop("Error: 'opts' has to be numeric!")
  }
```

Similarly, these checks for missing values should be rewritten:

```r
## check for missings
  if (anyNA(data %>% dplyr::select(., {{ opts }}))) {
    stop("Error: 'opts' contains NAs!")
  }

  # choice
  ## check for missing
  if (base::anyNA(data %>% dplyr::select(., {{ choice }}))) {
    stop("Error: 'choice' contains NAs!")
  }

```

```r
opts_contains_na <- any_var_na(data, {{ opts }} )
choice_contains_na <- any_var_na(data, {{ choice }} )

if (opts_contains_na) {
  stop("Error: 'opts' contains NAs!")
}
if (choice_contains_na) {
  stop("Error: 'choice' contains NAs!")
}
```

Even better would be wrapping these up into functions:

```r
check_if_missing <- function(data, vars, arg){
  vars_have_missings <- any_var_na(data, {{ vars }} )
  if (vars_have_missings){
    stop(sprintf("%s contains NAs!", arg))
  }
}

check_if_missing(data, {{ opts }}, "opts")
check_if_missing(data, {{ choice }}, "choice")
```

This check on choices here could be better expressed as error checking functions, so this:

```r
  ## check for str
  choi <- data %>%
    dplyr::select(., {{ choice }}) %>%
    base::colnames()

  if (!base::is.numeric(data[[choi]])) {
    stop("Error: 'choice' has to be numeric!")
  }
```

Would be:

```r
check_if_cols_numeric(data, choice)
```

A very similar check is done above for whether "opts" is numeric - so this could be a function that is used across a few different parts of the package.

See https://github.com/r-lib/rlang/blob/main/R/standalone-types-check.R for some possible existing helper functions. These checkers can be inserted into code with https://usethis.r-lib.org/reference/use_standalone.html

You do not need to use `return` in your code, as a function will return the last thing that is performed. You are welcome to use it, but I believe that the data frame that gets computed inside return should happen as a separate step that is assigned into a variable. This makes the code easier to debug, as otherwise you need to step into the `return` code, which is actually a bit tricky.

So instead of:

```r
  return(data %>%
    dplyr::mutate(
      # store column index with highest utility
      pred = base::max.col(dplyr::pick({{ opts }})),
      buy = base::ifelse({{ choice }} != base::match(
        data %>% dplyr::select(., {{ none }}) %>% colnames(),
        data %>% dplyr::select(., {{ opts }}) %>% colnames()
      ), 1, 2), # dichotomies actual choice (1 = prod, 2 = none)
      pred = base::ifelse(pred != base::match(
        data %>% dplyr::select(., {{ none }}) %>% colnames(),
        data %>% dplyr::select(., {{ opts }}) %>% colnames()
      ), 1, 2) # dichotomies pred choice (1 = prod, 2 = none)
    ) %>%
    dplyr::group_by(pick({{ group }})) %>% # potential grouping variable
    dplyr::summarise(
      accuracy = 100 * base::mean(buy == pred) # calculate accuracy
    ))
```

I suggest

```r
accuracy_data <- data %>%
    dplyr::mutate(
      # store column index with highest utility
      pred = base::max.col(dplyr::pick({{ opts }})),
      buy = base::ifelse({{ choice }} != base::match(
        data %>% dplyr::select(., {{ none }}) %>% colnames(),
        data %>% dplyr::select(., {{ opts }}) %>% colnames()
      ), 1, 2), # dichotomies actual choice (1 = prod, 2 = none)
      pred = base::ifelse(pred != base::match(
        data %>% dplyr::select(., {{ none }}) %>% colnames(),
        data %>% dplyr::select(., {{ opts }}) %>% colnames()
      ), 1, 2) # dichotomies pred choice (1 = prod, 2 = none)
    ) %>%
    dplyr::group_by(pick({{ group }})) %>% # potential grouping variable
    dplyr::summarise(
      accuracy = 100 * base::mean(buy == pred) # calculate accuracy
    )
  return(accuracy_data)
```

I also think that you should convert this code:

```r
 base::ifelse({{ choice }} != base::match(
        data %>% dplyr::select(., {{ none }}) %>% colnames(),
        data %>% dplyr::select(., {{ opts }}) %>% colnames()
      ), 1, 2), # dichotomies actual choice (1 = prod, 2 = none)
```

Into a function since you use it twice. For example:

```r
colnames_match <- function(data, set_one, set_two){
  names_none <- dplyr::select(data, {{ set_one }}) |> names()
  names_opts <- dplyr::select(data, {{ set_two }}) |> names()
  match(names_none, names_opts)
}

dichotimise_choice <- function(data, test){
  ifelse(
    test = test
    # dichotomies actual choice (1 = prod, 2 = none)
    yes = 1,
    no = 2
}

...

mutate(
... = dichotimise_choice({{ choice }} != colnames_match(none, opts))
... = dichotimise_choice(colnames_match(none, opts))
)

```

## an nvars helper function

There is a lot of this pattern in the code:

```r
if (base::length(data %>% dplyr::select(., {{ opts }})) == 1) {
# error message
}
```

As mentioned above, there are places where `missing()` would be a more precise function to use.

But if you do need to determine how many columns there are, then I believe a function like `n_vars`, defined as below would be helpful:

```r
n_vars <- function(data, cols){
  data |>
    dplyr::select( {{ cols }} ) |>
    ncol()
}
```

It could then be used like so:

```r
n_cols_none <- n_vars(data, {{none}})
if (n_cols_none > 1){
# error message
}
```

## using `base::fn`

You do not need to namespace `base::` calls, since the `base` package is always provided with R. I suggest removing all instances of this as it unnecessarily clutters the code.

# att_imp.R

## Repeated patterns

As suggested above, this pattern is repeated a lot:

```r
if (base::anyNA(data %>% dplyr::select(., {{ group }}))) {
```

I suggest wrapping this up in a function

## lots of error/warning checking

It's great that the author is writing a lot of warnings and checks, but I think that these could be wrapped up in other checking functions, as these checking functions are obscuring the intention of the function. I wrote about this here, it might be of interest: https://www.njtierney.com/post/2023/12/06/long-errors-smell/

## using a for loop to grow a vector

Code like

```r
 att <- base::length(attrib)

  new <- c()

  helper <- 1
  for (i in 1:att) {
    data[[base::paste0("att_imp_", i)]] <- 0

    vars <- attrib[[i]]

    new <- c(new, base::paste0("att_imp_", i))

```

Can be dangerous to memory as `new` is not assigned a size or length. This could be done instead with

```r
vector(mode = "character", length = length(attrib))
```

However, the larger issue here is that this code does not need to use a for loop - code like

```r
data[[base::paste0("att_imp_", i)]] <- 0
```

Can instead be

```r
new_col_names <- paste0("att_imp_", 1:10)
data[[new_col_names]] <- 0
```

Which is much more efficient.

## using return statements that evaluate code

Similar to above, I suggest changing this pattern:

```r
return(
data |> function(args) |> function()
)
```

to be

```r
data_to_return <- data |> function(args) |> function()
return(data_to_return)
```

# createHOT.R

## Checking/defensive code

Code like

```r
  # test numeric input for coding
  if (!(base::is.null(coding))) {
    for (i in 1:base::length(coding)) {
      if (!(base::is.numeric(coding[i]))) {
        base::stop("Error: 'coding' only can have numeric input!")
      }
    }
  }
```

Does not need a for loop, and could be expressed as:

```r
  if (!is.null(coding)) {
    coding_numeric <- is.numeric(coding)
    if (!any(coding_numeric)){
        stop("Error: 'coding' only can have numeric input!")
    }
  }
```

And similarly for the other for:

```r
  if (!(base::is.null(coding))) {
    if (base::any(coding == 1) | base::any(coding == 2)) {
      for (q in 1:base::length(prod.levels)) {
        if (base::any(is.character(prod.levels[[q]]))) {
          base::stop(
            "Error: Input 'prod.levels' has to be numeric if linear or ",
            "piecewise coded variables are included. Please use column indexes ",
            "instead of column names if you specify columns!"
          )
        }
      }
    }
  }
```

## use match.arg

The code:

```r

  # test whether method is correctly specified
  if ((method != "ACBC") & (method != "CBC") & (method != "MaxDiff")) {
    base::stop(
      "Error: Please choose one of the supported methods: 'MaxDiff',",
      " 'ACBC', 'CBC'!"
    )
  }
```

can be replaced with:

```r
method <- match.arg(method, c("ACBC", "CBC", "MaxDiff"))
```

The `match.arg` function handles this for you - try running:

```r
match.arg("example_arg", c("a", "b", "c"))
```

## deeply nested if/for loops

This code

```r
  # test variables of prod.levels
  if (!(base::is.null(prod.levels))) {
    for (tt in 1:base::length(prod.levels)) {
      lng <- base::length(prod.levels[[tt]])
      if (lng == 1) {
        if (!base::is.na(prod.levels[[tt]])) {
          var <- prod.levels[[tt]]
          if (!(base::is.numeric(data[[var]]))) {
            base::stop(
              "Error: Variables included in 'prod.levels' ",
              "have to be numeric!"
            )
          }
        }
      }

      if (lng > 1) {
        for (lng_lev in 1:lng) {
          if (coding[lng_lev] != 1 & coding[lng_lev] != 2) {
            if (!base::is.na(prod.levels[[tt]][lng_lev])) {
              var <- prod.levels[[tt]][lng_lev]
              if (!(base::is.numeric(data[[var]]))) {
                base::stop(
                  "Error: Variables included in 'prod.levels' ",
                  "have to be numeric!"
                )
              }
            }
          }
        }
      }
    }
  }
```

Is very hard to error check - I believe a lot of the complexity here could be managed by using vectorised functions in R and indexing appropriately. For example, you don't need to do a for loop to get each length, you can use `lengths`, which is a base R function - see for example,

```r
length(iris)
lengths(iris)
```

# general comments

For all code files, see above comments about:

- wrapping error checking code into functions
- avoiding for loops that can instead use vectorised R functions
- only use return to return a single object, don't wrap return around longer code blocks

## documentation

The authors can use markdown inside their roxygen code. This means they can use write:

```
`accuracy` is one of the 5 metrics of the confusion matrix
```

instead of:

```
\code{accuracy} is one of the 5 metrics of the confusion matrix
```

See https://github.com/r-lib/roxygen2md to assist doing this.

## Consider using `across` instead of `dplyr::mutate_at`

mutate_at is now superceded, it might be better for the long term of your package to look into the newer methods, but up to you!

## Consider using switch instead of multiple if statements

instead of:

```r
if (condition == "one"){
 # many lines of code
 # ...
}

if (condition == "two"){
# many lines of code
 # ...
}

if (condition == "three"){
# many lines of code
 # ...
}
```

You can use `switch`

```r
data_output <- switch(
  condition,
  one = function_for_one(data),
  two = function_for_two(data),
  three = function_for_three(data)
  )

  return(data_output()
```

[…]
