+++ paper/paper.bib
@@ -329,3 +329,37 @@ analysis of Best-Worst Surveys},
 	date = {2023},
 	url = {https://CRAN.R-project.org/package=dplyr}
 }
+
+@article{schramm2024,
+	title = {Incentive alignment in anchored MaxDiff yields superior predictive validity},
+	author = {Schramm, Joshua Benjamin and Lichters, Marcel},
+	year = {2024},
+	month = {01},
+	date = {2024-01-11},
+	journal = {Marketing Letters},
+	doi = {10.1007/s11002-023-09714-2},
+	url = {http://dx.doi.org/10.1007/s11002-023-09714-2},
+	langid = {en}
+}
+
+
+@article{magrittr,
+	title = {magrittr: A Forward-Pipe Operator for R},
+	author = {Bache, Stefan Milton and Wickham, Hadley},
+	year = {2022},
+	date = {2022},
+	url = {https://CRAN.R-project.org/package=magrittr}
+}
+
+@article{steiner2018,
+	title = {A User's Guide to the Galaxy of Conjoint Analysis and Compositional Preference Measurement},
+	author = {Steiner, Michael and {Meißner}, Martin},
+	year = {2018},
+	date = {2018},
+	journal = {Marketing ZFP},
+	pages = {3--25},
+	volume = {40},
+	number = {2},
+	doi = {10.15358/0344-1369-2018-2-3},
+	url = {http://dx.doi.org/10.15358/0344-1369-2018-2-3}
+}
