+++ paper/paper.md
@@ -1,236 +1,0 @@
----
-title: 'validateHOT - an R package for holdout task validation and market simulations'
-tags:
-  - R
-  - MaxDiff
-  - Conjoint Analysis
-  - Market Simulations
-  - Predictive Validity
-authors:
-  - name: Joshua Benjamin Schramm
-    orcid: 0000-0001-5602-4632
-    corresponding: True
-    affiliation: 1
-  - name: Marcel Lichters
-    orcid: 0000-0002-3710-2292
-    corresponding: FALSE
-    affiliation: 1    
-affiliations:
- - name: Otto von Guericke University of Magdeburg, Germany
-   index: 1
-citation_author: Schramm & Lichters
-date: 07 July
-year: 2024
-bibliography: paper.bib
-link-citations: true
-output: rticles::joss_article
-journal: JOSS
----
-
-\newcommand{\colcod}[1]{\texttt{\color{purple}#1}}
-
-
-
-# Summary
-
-validateHOT is an R package that provides functions to both validate a validation/holdout task and run market simulations for results obtained in a (adaptive) choice-based conjoint analysis (hereafter ACBC and CBC, respectively) and maximum difference scaling (hereafter MaxDiff) using, for example, ChoiceModelR [@ChoiceModelR] or Sawtooth's Lighthouse Studio.
-
-# Statement of need
-
-Preference measurement techniques' (e.g., (A)CBC or MaxDiff) aim is to predict behavior [@green1990]. Hence, it is essential to ensure that the collected data is valid and predicts outside tasks (i.e., the model has external validity) well. The easiest way for testing validity is by including validation tasks [e.g., @Orme2015], which are fixed tasks (i.e., same across participants) and not used for estimating the part-worth utilities (raw logit utilities) in hierarchical Bayes (HB) estimation. Despite their importance, practitioners don't always include them [@yang2018]. This is unsatisfactory given the fact that the model is used to estimate market shares which poses the basis for relevant marketing decisions.
-
-validateHOT combines both validation and market simulation in one package and has three key advantages, it (1) helps opting for the best model and (2) runs relevant market simulations that help finding the right product combinations or assortments, and (3) is an open source tool which helps especially researchers reporting accompanied scripts for their research papers.
-
-# State of the field in R
-
-Other packages provide functions to calculate validation metrics, however, these are not always specified for individual part-worth utilities. Metrics [@Metrics], for example, provide functions to run validation metrics such as *mean absolute error*, *root mean squared error*, or the five metrics of the confusion matrix. However, to get the output of, for example, Sawtooth Software or ChoiceModelR [@ChoiceModelR] into the right format, the user needs some data wrangling. The package conjoint [@conjoint] provides functions that are most similar to validateHOT's ones, but no validation functions are included and the package focuses on classical conjoint analysis, thus it is limited when applying more common conjoint methods. logitr [@logitr] provides market simulations tools, however, no validation metrics such as mean hit probability or hit rate. \autoref{comparison} shows a comparison of validateHOT's functions with current R packages. To the best of our knowledge, a package that converts raw utility scores into validation metrics or running a variety of marketing simulations (especially TURF) is missing.
-
-```{=tex}
-\begin{figure}[h]
-  \includegraphics{figures/FunctionComparison.png}
[…]
