+++ paper/paper.Rmd
@@ -1,64 +1,65 @@
 ---
-title: 'validateHOT - an R package for holdout task validation and market simulations'
+title: "validateHOT - an R package for validating validation tasks and choice modelling tools"
 tags:
-  - R
-  - MaxDiff
-  - Conjoint Analysis
-  - Market Simulations
-  - Predictive Validity
-authors:
-  - name: Joshua Benjamin Schramm
-    orcid: 0000-0001-5602-4632
-    corresponding: True
-    affiliation: 1
-  - name: Marcel Lichters
-    orcid: 0000-0002-3710-2292
-    corresponding: FALSE
-    affiliation: 1    
-affiliations:
- - name: Otto von Guericke University of Magdeburg, Germany
-   index: 1
+- R
+- MaxDiff
+- Conjoint Analysis
+- Market Simulations
+- Predictive Validity
+date: "08 January"
+output:
+  pdf_document:
+    df_print: paged
 citation_author: Schramm & Lichters
-date: 07 July
-year: 2024
+authors:
+- name: Joshua Benjamin Schramm
+  orcid: "0000-0001-5602-4632"
+  corresponding: true
+  affiliation: 1
+- name: Marcel Lichters
+  orcid: "0000-0002-3710-2292"
+  corresponding: false
+  affiliation: 1
+year: 2025
 bibliography: paper.bib
 link-citations: true
-output: pdf_document
+affiliations:
+- name: Otto von Guericke University of Magdeburg, Germany
+  index: 1
 ---
-
-\newcommand{\colcod}[1]{\texttt{\color{purple}#1}}
 
 ```{r include=FALSE}
 library(validateHOT)
-library(tidyverse)
+library(magrittr)
 ```
 
 # Summary
 
-validateHOT is an R package that provides functions to both validate a validation/holdout task and run market simulations for results obtained in a (adaptive) choice-based conjoint analysis (hereafter ACBC and CBC, respectively) and maximum difference scaling (hereafter MaxDiff) using, for example, ChoiceModelR [@ChoiceModelR] or Sawtooth's Lighthouse Studio.
+validateHOT is an R package that provides functions for preference measurement techniques like (adaptive) choice-based conjoint analyses (hereafter (A)CBC) and maximum difference scaling (hereafter MaxDiff). More specifically with the package, users can validate validation tasks, perform market simulations, and rescaleg raw utility scores. The package works with data obtained using, for example, the ChoiceModelR package [@ChoiceModelR] or Sawtooth's Lighthouse Studio.[^1]
+
+[^1]: We refer to both validation and holdout tasks interchangeably.
 
 # Statement of need
 
-Preference measurement techniques' (e.g., (A)CBC or MaxDiff) aim is to predict behavior [@green1990]. Hence, it is essential to ensure that the collected data is valid and predicts outside tasks (i.e., the model has external validity) well. The easiest way for testing validity is by including validation tasks [e.g., @Orme2015], which are fixed tasks (i.e., same across participants) and not used for estimating the part-worth utilities (raw logit utilities) in hierarchical Bayes (HB) estimation. Despite their importance, practitioners don't always include them [@yang2018]. This is unsatisfactory given the fact that the model is used to estimate market shares which poses the basis for relevant marketing decisions.
+Researchers and practitioners use preference measurement techniques for many reasons, for example, simulating markets or to determine the importance of attributes to name a few [@steiner2018]. Their ultimate goal is to predict future behavior [@green1990]. In order to predict accurately and make the right decision, it is essential to ensure that the collected data is valid. One way to test the model's validity is by including validation tasks [e.g., @Orme2015], which are usually fixed tasks (i.e., same across participants) and excluded for utility estimation in hierarchical Bayes (HB) estimation.
 
[…]
