+++ tests/testthat/test-CBC_lin.R
@@ -1,6 +1,6 @@
-data("cbc_linear")
+df <- validateHOT::cbc_linear
 
 test_that("Structure of df", {
-  expect_equal(nrow(cbc_linear), 105)
-  expect_equal(ncol(cbc_linear), 36)
+  expect_equal(nrow(df), 105)
+  expect_equal(ncol(df), 36)
 })
