+++ tests/testthat/test-MaxDiff.R
@@ -1,6 +1,6 @@
-data(maxdiff)
+df <- validateHOT::maxdiff
 
 test_that("Structure of df", {
-  expect_equal(nrow(maxdiff), 118)
-  expect_equal(ncol(maxdiff), 20)
+  expect_equal(nrow(df), 118)
+  expect_equal(ncol(df), 20)
 })
