+++ tests/testthat/test-ACBC.R
@@ -1,6 +1,6 @@
-data(acbc)
+df <- validateHOT::acbc
 
 test_that("Structure of df", {
-  expect_equal(nrow(acbc), 110)
-  expect_equal(ncol(acbc), 37)
+  expect_equal(nrow(df), 110)
+  expect_equal(ncol(df), 37)
 })
