+++ tests/testthat/test-ACBC_interpolate.R
@@ -1,6 +1,6 @@
-data(acbc_interpolate)
+df <- validateHOT::acbc_interpolate
 
 test_that("Structure of df", {
-  expect_equal(nrow(acbc_interpolate), 110)
-  expect_equal(ncol(acbc_interpolate), 39)
+  expect_equal(nrow(df), 110)
+  expect_equal(ncol(df), 39)
 })
