+++ man/ACBC_interpolate.Rd
@@ -12,7 +12,7 @@ A data frame with 110 rows and 39 variables
 \url{https://osf.io/m5x3a/}
 }
 \usage{
-data(acbc_interpolate)
+acbc_interpolate
 }
 \description{
 A data set with raw scores export for an adaptive choice-based conjoint
@@ -70,6 +70,7 @@ price level 5}
 }
 }
 \examples{
+
 # create total utility of validation tasks
 
 prod1 <- c(3, 10, 9, 12, 15, 19, 23, 31, 248.55)
