### 2025-01-10 ba32fb0
updated workflow and paper - 250110
