### 2025-01-10 830fa26
updated manuals for functions - 250110
