+++ R/create_hot.R
@@ -1,10 +1,10 @@
 #' @title  Function to create total utilities for validation task alternatives
 #' or market scenario
 #'
-#' @param data A data frame with all relevant variables.
-#' @param id A vector of unique identifier in `data`.
-#' @param none An optional vector to specify `none`
-#' alternative in `data`.
+#' @param data A data.frame object.
+#' @param id Column name of the unique identifier.
+#' @param none An optional vector to specify column name of a `none`
+#' alternative.
 #' @param prod.levels A list to define the attribute levels of the
 #' alternatives (`prod`). If linear-coded or piecewise-coded
 #' attributes are included, column indexes are required for the input.
@@ -15,41 +15,42 @@
 #' Only has to be specified for the attributes that are coded as `1` (linear)
 #' or `2` (piecewise).
 #' @param piece.p A nested list, with a list for each of the piecewise-coded
-#' variables. List for a piecewise-cded attribute must be the columns that
+#' variables. List for a piecewise-coded attribute must be the columns that
 #' shoud be the lower and the upper level that should be used for interpolating.
 #' @param lin.p A vector to specify column index or column names of
 #' linear coded variables.
 #' @param coding A vector to define attributes coding, `0` = part-worth
 #' coding,`1` = linear coding, `2` = piecewise coding; please make sure to code
 #' linear price of ACBC as piecewise. For more details, see the example
-#' provided below.
+#' provided below. If you want to treat a part-worth coded variable
+#' continuously, use the code `2` for this variable and provide the values in
+#' `interpolate.levels` accordingly.
 #' @param method A character to specify the `method` of your study.
 #' `method` has to be one of the following: `"maxdiff"`, `"cbc"`, or `"acbc"`.
-#' @param varskeep A vector specifying variables that should be kept
-#' in the data frame.
+#' @param varskeep A vector specifying column names of the variables that should
+#' be kept in the data frame.
 #' @param choice Actual choice in the validation task. Leave empty for
 #' specifying market scenario (warning will be displayed, however).
 #'
 #' @details
 #' To test the validation metrics of a validation task or to run a
 #' market scenario, the scenario first has to be created by summing up
-#' the alternatives part-worth utilities. This is done with the `create_hot()`
+#' the alternatives raw utilities. This is done with the `create_hot()`
 #' function.  Make sure you upload the raw utilities of your study (either
 #' from Sawtooth Software or ChoiceModelR, Sermas, 2022).
 #' The function then creates the alternatives' utility based on the additive
 #' utility model (Rao, 2014, p. 82). If you are working with alternative
 #' specific-designs, insert `NA` if attribute is not specified.
 #'
-#' `data` has to be a data frame with raw scores of the attribute
+#' `data` must to be a data.frame object with raw scores of the attribute
 #' levels.
 #'
 #' `id` has to be the column index or column name of the id (unique for each
 #' participant) in data frame.
 #'
-#' `none` to specify a `none` alternative if it is included in validation task.
-#' Specify column index or column name of `none` alternative. If there was no
-#' `none` alternative in the validation task, leave it empty.
-#'
+#' `none` to specify variable name of the `none` alternative if it is included
+#' in the validation task. Leave it empty, if there was no `none` alternative
+#' included.
 #'
 #' `prod.levels` specifies the attribute levels for each alternative.
 #' Input for `prod.levels` has to be a list. If values for one attribute are
@@ -59,7 +60,8 @@
 #' specified.
 #'
 #' `interpolate.levels` is required in case interpolating is used (only if
[…]
