+++ R/turf.R
@@ -22,7 +22,7 @@
 #'
 #' @details
 #'
-#' `data` a data frame including the alternatives that should be tested.
+#' `data` a data.frame object including the alternatives that should be tested.
 #'
 #' `opts` to specify the different alternatives in the product assortment that
 #' should be considered. Input of `opts` has to be column name of variables
@@ -123,6 +123,7 @@ turf <- function(data,
                  fixed = NULL,
                  prohib = NULL,
                  approach = c("thres", "fc")) {
+
   # check for missing arguments ------------------------------------------------
   if (missing(opts)) {
     stop('Error: argument "opts" must be provided.')
