### 2025-01-11 35cd9f4
updated readme and vignette + took care of rmcd checks issues - 250111
