+++ paper/paper.bib
@@ -20,6 +20,8 @@
  title = {Including Holdout Choice Tasks in Conjoint Studies},
  file = {Orme 2015 - Including Holdout Choice Tasks:Attachments/Orme 2015 - Including Holdout Choice Tasks.pdf:application/pdf}
 }
+
+
 @article{yang2018,
 	title = {Attention, Information Processing, and Choice in Incentive-Aligned Choice Experiments},
 	author = {Yang, {Liu (Cathy)} and Toubia, Olivier and de Jong, Martijn G.},
