### 2025-01-11 26c5c10
adjusted paper.md - 250111
