+++ paper/paper.bib
@@ -365,3 +365,13 @@ analysis of Best-Worst Surveys},
 	doi = {10.15358/0344-1369-2018-2-3},
 	url = {http://dx.doi.org/10.15358/0344-1369-2018-2-3}
 }
+
+
+@misc{sawtooth2024,
+ author = {{Sawtooth Software Inc.}},
+ year = {2024},
+ title = {Lighthouse Studio 9},
+ url = {https://www.sawtoothsoftware.com/products/online-surveys},
+ address = {Sequim, WA.},
+ publisher = {{Sawtooth Software Inc}}
+}
