+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Type: Package
 Package: validateHOT
-Title: validateHOT - An R Package to validate holdout tasks and run market simulations 
+Title: validateHOT - an R package for validating validation tasks and choice modelling tools 
 Version: 1.0.4
 Authors@R: c(person("Joshua", "Schramm", email = "joshua.schramm@mail.de", role = c("aut", "cre"),
     comment = c(ORCID = "0000-0001-5602-4632")),
