### 2025-01-11 39ddbd3
updated dependencies & vignette - 250111
