+++ DESCRIPTION
@@ -24,7 +24,7 @@ Depends:
 Imports: 
     cli (>= 3.6.3),
     dplyr (>= 1.1.4),
-    fastDummies (>= 1.7.3),
+    fastDummies (>= 1.7.4),
     magrittr (>= 2.0.3),
     rlang (>= 1.1.4),
     scales (>= 1.3.0),
@@ -34,11 +34,11 @@ Imports:
     tidyselect (>= 1.2.1),
     utils (>= 4.3.2)
 Suggests: 
-    knitr (>= 1.47),
-    labelled (>= 2.13.0),
+    knitr (>= 1.49),
+    labelled (>= 2.14.0),
     Metrics (>= 0.1.4),
-    rmarkdown (>= 2.27),
-    testthat (>= 3.2.1.1)
+    rmarkdown (>= 2.29),
+    testthat (>= 3.2.2)
 VignetteBuilder: 
     knitr,
     dplyr,
