+++ paper/paper.md
@@ -1,6 +1,6 @@
 ---
-title: "validateHOT - an R package for validating validation tasks and choice modelling
-  tools"
+title: "validateHOT - an R package for the analysis of holdout/validation 
+tasks and other choice modeling tools"
 tags:
 - R
 - MaxDiff
@@ -19,41 +19,41 @@ authors:
 affiliations:
 - name: Otto von Guericke University of Magdeburg, Germany
   index: 1
-date: 11 January 2025
+date: 16 January 2025
 bibliography: paper.bib
 output: pdf_document
 ---
 
 # Summary
 
-validateHOT is an R package that provides functions for preference measurement techniques like (adaptive) choice-based conjoint analyses (hereafter (A)CBC) and maximum difference scaling (hereafter MaxDiff). More specifically with the package, users can validate validation tasks, perform market simulations, and rescale raw utility scores. The package works with data obtained using, for example, the ChoiceModelR package [@ChoiceModelR] or Sawtooth's Lighthouse Studio [@sawtooth2024].[^1]
+validateHOT is an R package that provides functions for preference measurement techniques such as (adaptive) choice-based conjoint analyses (hereafter CBC and ACBC, respectively) and maximum difference scaling (hereafter MaxDiff). Specifically, the package allows users to validate validation tasks, perform market simulations, and rescale raw utility scores. It is compatible with data obtained using, for example, the ChoiceModelR package [@ChoiceModelR] or Sawtooth Software's Lighthouse Studio [@sawtooth2024].[^1]
 
 [^1]: We refer to both validation and holdout tasks interchangeably.
 
 # Statement of need
 
-Researchers and practitioners use preference measurement techniques for many reasons, for example, simulating markets or to determine the importance of attributes [@steiner2018]. Their ultimate goal is to predict future behavior [@green1990]. In order to predict accurately and make the right decisions, it is essential to ensure that the collected data is valid. One way to test the model's validity is by including validation tasks [e.g., @Orme2015], which are usually fixed tasks (i.e., same across participants) and excluded for utility estimation in hierarchical Bayes (HB) estimation.
+Researchers and practitioners use preference measurement techniques for various research purposes, such as simulating markets or determining the importance of attributes when making choices [@steiner2018]. The ultimate goal is to predict future behavior [@green1990]. To make accurate predictions and informed decisions, it is crucial to ensure that the collected data is valid. A prominent way to test a model's validity is the analysis of validation tasks [e.g., @Orme2015], which are typically fixed tasks (i.e., same across participants) and excluded for utility estimation.
 
-The validateHOT provides the relevant tools for these steps: (1) it assesses the model's validity, (2) runs relevant market simulations, (3) converts raw utilities scores into scores that are easy to interpret. Finally, it is an open source tool helping researchers reporting accompanied scripts for their research papers.
+The validateHOT package provides the necessary tools for the aforementioned applications: (1) assessing the model's validity, (2) performing market simulations, (3) converting raw utility scores. Finally, as an open-source tool, it helps academics to report analysis scripts in scientific journal articles, fostering reproducibility.
 
-# State of the field in R
+# State of the field
 
[…]
