+++ R/MHP.R
@@ -1,4 +1,4 @@
-#' Function to calculate Mean Hit Probability (MHP) of validation task
+#' Function to calculate the mean hit probability of a validation task
 #'
 #' @description
 #' `mhp()` calculates the mean hit probability of a validation
@@ -13,20 +13,20 @@
 #' task.
 #'
 #' @details
-#' Mean hit probability (MHP) calculates the averaged hit probability of
+#' Mean hit probability calculates the averaged hit probability of
 #' participants actual choices in the validation task.
 #'
 #' `data` a `data.frame` object including the alternatives shown in the
 #' validation task. It can be created using the `create_hot()` function.
 #'
 #' `group` optional grouping variable(s) to display results by group(s).
-#' different groups. Has to be the column name(s) of variables in `data`.
+#' Has to be the column name(s) of variables in `data`.
 #'
 #' `opts` to specify the different alternatives in the
 #' validation task (also includes the `none` alternative).
 #'
-#' `choice` to specify the column of actual choice in the validation
-#' task. The input of `choice` has to be the column name of actual choice.
+#' `choice` to specify the column of the actual choice in the validation
+#' task. The input of `choice` has to be the column name of the actual choice.
 #'
 #' Output will display both mean hit probability and its corresponding standard
 #' error (both in percentage).
