+++ R/KL.R
@@ -18,7 +18,7 @@
 #' @details
 #' Kullback-Leibler-Divergence calculates the divergence between the actual
 #' choice distribution and the predicted choice distribution (Ding et al., 2011;
-#' Drost, 2018). Currently `kl()` provides the deviation measured based on
+#' Drost, 2018). Currently, `kl()` provides the deviation measured based on
 #' \eqn{log} and \eqn{log{_2}} algorithm. \eqn{log} is set as default.
 #'
 #' Due to Kullback-Leibler divergence's asymmetry, the output provides both
@@ -28,14 +28,14 @@
 #' `data` a `data.frame` object including the alternatives shown in the
 #' validation task. It can be created using the `create_hot()` function.
 #'
-#' `group` optional grouping variable(s) to calculate results by group(s).
-#' different groups. Has to be the column name(s) of variables in `data`.
+#' `group` optional grouping variable(s) to display results by group(s).
+#' Has to be the column name(s) of variables in `data`.
 #'
 #' `opts` to specify the different alternatives in the
 #' validation task (also includes the `none` alternative).
 #'
-#' `choice` to specify the column of actual choice in the validation
-#' task. The input of `choice` has to be the column name of actual choice.
+#' `choice` to specify the column of the actual choice in the validation
+#' task. The input of `choice` has to be the column name of the actual choice.
 #'
 #' `epsilon` numeric input in case of `0` in the numerator or
 #' denominator. `0` then will be replaced by `epsilon`. The default value
