### 2025-01-22 b97fe19
proofread README and vignette - 250122
