+++ NEWS.md
@@ -1,12 +1,16 @@
 # validateHOT 1.0.4
 
-* Implemtend the suggestions from the JOSS review process, which makes functions faster
+* Implemented the suggestions from the JOSS review process, which makes functions faster
 * changed name from createHOT() to create_hot()
 * changed piece.p to a nested list to allow for multiple piecewise-coded attributes
-* added examples for parth-worth coded attributes but later treat this attribute continuously
+* added examples for part-worth coded attributes but later treat this attribute continuously
 * added turf_ladder() function
 * bug fixes
 * adjusted DESCRIPTION file title slightly
+
+# validateHOT 1.0.3
+
+* got rid of the base:: calls in all functions
 
 # validateHOT 1.0.2
 
