+++ README.Rmd
@@ -27,8 +27,8 @@ knitr::opts_chunk$set(
 
 <!-- badges: end -->
 
-validateHOT is a package for conjoint and MaxDiff users. It provides
-functions to evaluate validation tasks, run market simulations, and
+validateHOT is a package for preference measurement techniques. It provides
+functions to evaluate validation tasks, perform market simulations, and
 convert raw utility estimates into scores that are easier to interpret.
 All three components are key functions for preference measurement
 techniques such as choice-based conjoint (CBC), adaptive choice-based
@@ -38,7 +38,7 @@ Software](https://sawtoothsoftware.com/) community who would like to
 report their analysis in *R* for open science purposes. In addition, it
 is compatible with other packages, for example, the ChoiceModelR package
 [@ChoiceModelR]. Further, the validateHOT package is valuable for
-practitioners, who would like to conduct the analyses using open-source
+practitioners, who would like to conduct the analyses using an open-source
 software.
 
 Researchers and practitioners use preference measurement techniques for
@@ -51,7 +51,7 @@ estimation of the utility scores. Including validation tasks is highly
 recommended [@Orme2015; @rao2014]. They do not only verify data validity
 but can also help to test different models. The validateHOT package
 offers helpful tools, to facilitate these functions (i.e., validate a
-validation task, run market simulations, and communicate results of
+validation task, perform market simulations, and communicate results of
 preference measurement techniques) - all within an open-source tool.
 
 > The validateHOT package was primarily developed for use with Sawtooth
@@ -60,15 +60,11 @@ preference measurement techniques) - all within an open-source tool.
 > variables).
 
 👉🏾 <u>**What you need to provide**</u>: <br> After collecting your data
-and running your initial hierarchical Bayes model, the validateHOT
-package allows you to test how well your model predicts choices in the
-validation task, perform market simulations or convert your data into
-scores that are easier to interpret. To use the validateHOT package, you
-need to import in your raw utility scores. If you plan to validate a
-validation task you also need to provide the actual choice made in this
-task. We provide a short tutorial in this markdown. For a more
-comprehensive tutorial, please see the vignette provided with the
-validateHOT package
+and running your initial hierarchical Bayes model, you need to import your raw 
+utility scores. If you plan to validate a validation task you also need to 
+provide the actual choice made in this task. We provide a short tutorial in 
+this markdown. For a more comprehensive tutorial, please see the vignette 
+provided with the validateHOT package
 (`vignette("validateHOT", package = "validateHOT")`).
 
 👈🏾 <u>**What you get**</u>:<br> The validateHOT package currently
@@ -85,8 +81,8 @@ provides functions for four key components:
 
 For the first three components, the `create_hot()` function is essential.
 This function calculates the total utilities for each alternative in the
-validation task and the market simulation.
-The `create_hot()` finction computes the total utility of each alternative
+validation task or the market scenario to be tested.
+The `create_hot()` function computes the total utility of each alternative
 based on the additive utility model [@rao2014, p. 82].
 
 ### Classical validation metrics
@@ -95,23 +91,25 @@ based on the additive utility model [@rao2014, p. 82].
     the validation task.
 -   `kl()`: Kullback-Leibler-Divergence calculates the divergence between
     the actual choice distribution and the predicted choice distribution
-    [@ding2011; @philentropy]. Due to the asymmetry of the Kullback-Leibler divergence, the output includes divergence both from predicted to observed and from observed to predicted. The validateHOT
[…]
