### 2025-01-30 #6708 @njtierney
Thank you for your response to my comments, @JoshSchramm94 ! I've had a look through your code and can see that you've implemented many changes, which I can appreciate would have taken a lot of time and effort.

One small thing, which is more of a "good to know about", but not something that I think you need to implement for this review, is patterns like this

```r
if (missing(method)) {
    method <- "stop"
  }
```

If you instead set the default value in function definition of `method` to `NULL`, then you can do this:

```r
my_function <- function(method = NULL){
  method <- method %||% "stop"
}
```

Which reads as "if method contains a value, use that, otherwise if it is NULL, give it "stop".

This, `%||%`, is called the "null pipe". It can be imported from `rlang`, or if you use default R 4.3.0 it is included in that version.

Just a small thing! But I think it's kind of neat.
