### 2025-01-30 #6708 @njtierney
I do still think that there are some opportunities to use `switch`, for example, https://github.com/JoshSchramm94/validateHOT/blob/main/R/MarkSim.R#L149-L159

But I can see that you've made a strong effort to put more things in functions, and I am really happy with the changes to the code overall.

Thanks again for all the changes you've made, I hope it's been an enjoyable process!
