+++ NAMESPACE
@@ -42,6 +42,7 @@ importFrom(dplyr,ungroup)
 importFrom(dplyr,vars)
 importFrom(fastDummies,dummy_cols)
 importFrom(magrittr,"%>%")
+importFrom(rlang,"%||%")
 importFrom(rlang,caller_arg)
 importFrom(rlang,caller_env)
 importFrom(stats,approx)
