+++ R/create_hot.R
@@ -475,11 +475,11 @@ create_hot <- function(data,
         if (min(linear_levels) == prod.levels[[q]][p]) {
           lower_break <- max(
             linear_levels[linear_levels <= prod.levels[[q]][p]]
-            )
+          )
 
           upper_break <- min(
             linear_levels[linear_levels > prod.levels[[q]][p]]
-            )
+          )
         }
 
         if (min(linear_levels) != prod.levels[[q]][p]) {
