+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: validateHOT
 Title: validateHOT - an R package for validating validation tasks and choice modelling tools 
-Version: 1.0.4
+Version: 1.0.5
 Authors@R: c(person("Joshua", "Schramm", email = "joshua.schramm@mail.de", role = c("aut", "cre"),
     comment = c(ORCID = "0000-0001-5602-4632")),
     person("Marcel", "Lichers", role = "ctb",
