### 2025-01-30 #6708 @JoshSchramm94
Thanks for your helpful comments @njtierney.

I upadted both the `kl()` and the `marksim()` functions in the way you suggested.
For the `marksim()` function, I set the default value of the `res` and `method` arguments to `NULL` and implemented the "null pipe" accordingly. For the `kl()` function, I did the same for the `epsilon` and the `log_base` arguments.

In addition, I added a new function in the `utils.R` file (`create_shares()`), in which I implemented the `switch()` function. This replaced the part you highlighted in your second comment.

I updated the Acknowledgments of the paper slighty and set the package version to 1.0.5.
I got one more question, how can I add you as a contributor? Or do I add the Reviewers to the `DESCRIPTION` file?

And yes, thank you so much for your comments, I was able to learn so much and I am extremely thankful for the time and effort the review team put into this process. Thank you so much for all these helpful advice. I really appreciate.
