### 2025-02-09 971a9a1
adjusted JOSS paper - 250209
