+++ paper/paper.md
@@ -40,9 +40,12 @@ The validateHOT package provides the necessary tools for the aforementioned appl
 
 # State of the field
 
-Other R packages offer functions to calculate validation metrics. However, these are not always tailored to the individual raw utilities extracted from preference measurement techniques. The Metrics package [@Metrics], for example, provides functions to run validation metrics such as mean absolute error (MAE) or metrics of the confusion matrix. However, converting outputs such as those from estimations using Sawtooth Software's Lighthouse Studio [@sawtooth2024] or the ChoiceModelR package [@ChoiceModelR] into the proper format requires some data wrangling. The conjoint package [@conjoint] offers functions that are similar to those in validateHOT but it lacks validation functions and focuses primarily on classical conjoint analysis. Thus, it is limited when applied to more common conjoint methods. The logitr package [@logitr] offers market simulation tools but does not include validation metrics such as mean hit probability or hit rate. \autoref{comparison} compares validateHOT's functions with those of other R packages. To the best of our knowledge, no package converts raw utility scores into validation metrics or running a variety of marketing simulations (especially Total Unduplicated Reach and Frequency (TURF) and TURF ladder).
+Other R packages offer functions to calculate validation metrics. However, these are not always tailored to the individual raw utilities extracted from preference measurement techniques. The Metrics package [@Metrics], for example, provides functions to run validation metrics such as mean absolute error (MAE) or metrics of the confusion matrix. However, converting outputs such as those from estimations using Sawtooth Software's Lighthouse Studio [@sawtooth2024] or the ChoiceModelR package [@ChoiceModelR] into the proper format requires some data wrangling. The conjoint package [@conjoint] offers functions that are similar to those in validateHOT but it lacks validation functions and focuses primarily on classical conjoint analysis. Thus, it is limited when applied to more common conjoint methods. The logitr package [@logitr] offers market simulation tools but does not include validation metrics such as mean hit probability or hit rate. \autoref{fig:comparison} compares validateHOT's functions with those of other R packages. To the best of our knowledge, no package converts raw utility scores into validation metrics or running a variety of marketing simulations (especially Total Unduplicated Reach and Frequency (TURF) and TURF ladder).
 
-<center>![Comparison of the functions of validateHOT with those of existing R packages](figures/functioncomparison.png){#comparison}</center>
+
+![Comparison of the functions of validateHOT with those of existing R packages\label{fig:comparison}](functioncomparison.png)
+
+
 
 We introduce validateHOT drawing on data estimated using Sawtooth Software Lighthouse Studio [@sawtooth2024]. However, it can also be used on results stemming from packages such as ChoiceModelR [@ChoiceModelR], bayesm [@bayesm], or STAN [@rstan], if used with similar settings.
 
@@ -51,7 +54,7 @@ We introduce validateHOT drawing on data estimated using Sawtooth Software Light
 validateHOT's functions can be categorized into four main components (see \autoref{tab:table1}). To convert the data in the correct format for most functions, we created the `create_hot()` function, which calculates each alternative's total utility in conjoint studies by applying the additive utility model.
 
 | Validation metrics | Confusion matrix | Market simulations | Rescaling scores |
-|:-----------------:|:----------------:|:-----------------:|:----------------:|
+|:----------------:|:----------------:|:----------------:|:----------------:|
 |     hitrate()      |    accuracy()    |    freqassort()    |    att_imp()     |
 |        kl()        |       f1()       |     marksim()      |  prob_scores()   |
 |       mae()        |   precision()    |      reach()       |    zc_diffs()    |
