### 2025-02-11 #6708 @jamesuanhoro
Yes, I am. Thank you.
