### 2025-02-13 #6708 @JoshSchramm94
@skanwal thanks a lot for getting back. Just a quick question, should I already get started with the authors tasks?
