### 2025-02-14 d53b223
adjusted paper slightly for JOSS - 250214
