### 2025-02-14 #6708 @JoshSchramm94
@skanwal we released the version [v1.0.5](https://github.com/JoshSchramm94/validateHOT/releases/tag/v1.0.5) and also archived the release on [zenodo](https://zenodo.org/records/14868050). The DOI is 10.5281/zenodo.14868050

If there is anything else to do, please let me know 😊
