### 2025-02-25 #6708 @JoshSchramm94
Hey @crvernon,
first of all, thanks a lot for handling the paper. Just a quick question: Should I proofread the paper again or adjust the missing DOIs? Thanks a lot in advance 😊
