### 2025-03-03 599b7e1
updated references for JOSS - 250303
