+++ paper/paper.bib
@@ -1,5 +1,5 @@
-﻿@article{green1990,
-	title = {Conjoint Analysis in Marketing: New Developments with Implications for Research and Practice},
+@article{green1990,
+	title = {{Conjoint Analysis in Marketing: New Developments with Implications for Research and Practice}},
 	author = {Green, Paul E. and Srinivasan, V.},
 	year = {1990},
 	month = {10},
@@ -17,7 +17,7 @@
 @misc{Orme2015,
  author = {Orme, Bryan K.},
  year = {2015},
- title = {Including Holdout Choice Tasks in Conjoint Studies},
+ title = {{Including Holdout Choice Tasks in Conjoint Studies}},
  file = {Orme 2015 - Including Holdout Choice Tasks:Attachments/Orme 2015 - Including Holdout Choice Tasks.pdf:application/pdf}
 }
 
@@ -37,11 +37,10 @@
 	langid = {en}
 }
 
-@article{Metrics,
+@Manual{Metrics,
 	title = {Metrics: Evaluation Metrics for Machine Learning},
 	author = {Hamner, Ben and Frasco, Michael},
 	year = {2018},
-	date = {2018},
 	url = {https://CRAN.R-project.org/package=Metrics}
 }
 
@@ -56,10 +55,10 @@
 }
 
 @article{ChoiceModelR,
-	title = {ChoiceModelR: Choice Modeling in R},
+	title = {ChoiceModelR: Choice {M}odeling in {R}},
 	author = {Sermas, Ryan},
 	year = {2022},
-	date = {2022},
+	doi = {10.32614/cran.package.choicemodelr},
 	url = {https://CRAN.R-project.org/package=ChoiceModelR}
 }
 
@@ -72,16 +71,18 @@ analysis of Best-Worst Surveys},
 }
 
 @article{logitr,
-	title = {{\textbraceleft}logitr{\textbraceright}: Fast Estimation of Multinomial and Mixed Logit Models with Preference Space and Willingness-to-Pay Space Utility Parameterizations},
-	author = {Helveston, John Paul},
-	year = {2023},
-	date = {2023},
-	volume = {105},
-	doi = {10.18637/jss.v105.i10}
+	title = {{logitr}: Fast Estimation of Multinomial and Mixed Logit Models with Preference Space and Willingness-to-Pay Space Utility Parameterizations},
+    author = {John Paul Helveston},
+    journal = {Journal of Statistical Software},
+    year = {2023},
+    volume = {105},
+    number = {10},
+    pages = {1--37},
+    doi = {10.18637/jss.v105.i10}
 }
 
 @article{bayesm,
-	title = {bayesm: Bayesian Inference for Marketing/Micro-Econometrics},
+	title = {{\lowercase{b}}ayesm: Bayesian Inference for Marketing/Micro-Econometrics},
 	author = {Rossi, Peter},
 	year = {2023},
 	date = {2023},
@@ -173,9 +174,9 @@ analysis of Best-Worst Surveys},
 @book{chrzan2019,
  author = {Chrzan, Keith and Orme, Bryan K.},
  year = {2019},
- title = {Applied MaxDiff: A practitioner's guide to best-worst scaling},
+ title = {{Applied MaxDiff: A Practitioner's Guide to Best-Worst Scaling}},
  address = {Provo, UT},
- publisher = {{Sawtooth Software}},
+ publisher = {{Sawtooth Software Inc.}},
  isbn = {9780999367728},
  file = {Chrzan, Orme 2019 - Applied MaxDiff:Attachments/Chrzan, Orme 2019 - Applied MaxDiff.pdf:application/pdf}
 }
@@ -270,7 +271,7 @@ analysis of Best-Worst Surveys},
[…]
