### 2025-03-03 #6708 @JoshSchramm94
:wave: @crvernon thanks for handling the paper, these are great news :blush:

I updated both accordingly

- [ ] LINE 90 and 113: The "r" representing the R programming language should be capitalized. You can do this by placing any characters you wish to maintain formatting of in the bib file in curly brackets.

- [ ] LINE 115: These capitalizations and the showing brackets need to be fixed as well.
